pyproject.toml
[project]
name = "eclass-mcp-server"
version = "0.1.0"
description = "An MCP server for the Open eClass platform by GUnet"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.5.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.2",
    "python-dotenv>=1.0.0",
]
[[project.authors]]
name = "CobuterMan"
email = "haidemenoss@gmail.com"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.scripts]
eclass-mcp-server = "eclass_mcp_server:main"
